backend/requirements.txt
fastapi
uvicorn
python-multipart
huggingface-hub
python-dotenv
opencv-python-headless
numpy
pillow
groq
tavily-python